+++ tests/cfo/scripts/list-market.js
@@ -4,7 +4,7 @@
 
 const utils = require("../tests/utils");
 const fs = require("fs");
-const anchor = require("@project-serum/anchor");
+const anchor = require("@coral-xyz/anchor");
 const provider = anchor.AnchorProvider.local();
 // hack so we don't have to update serum-common library
 // to the new AnchorProvider class and Provider interface