+++ tests/misc/tests/init-if-needed/init-if-needed.ts
@@ -1,5 +1,5 @@
-import * as anchor from "@project-serum/anchor";
-import { AnchorError, Program } from "@project-serum/anchor";
+import * as anchor from "@coral-xyz/anchor";
+import { AnchorError, Program } from "@coral-xyz/anchor";
 import { InitIfNeeded } from "../../target/types/init_if_needed";
 import { SystemProgram, LAMPORTS_PER_SOL } from "@solana/web3.js";
 import { expect } from "chai";