+++ ts/tests/src/spl/associated-token-account.ts
@@ -1,5 +1,5 @@
-import { splAssociatedTokenAccountProgram } from "@project-serum/spl-associated-token-account";
-import { splTokenProgram } from "@project-serum/spl-token";
+import { splAssociatedTokenAccountProgram } from "@coral-xyz/spl-associated-token-account";
+import { splTokenProgram } from "@coral-xyz/spl-token";
 import { Keypair, PublicKey, SystemProgram } from "@solana/web3.js";
 
 import { SPL_ATA_PROGRAM_ID, SPL_TOKEN_PROGRAM_ID } from "../constants";