+++ ts/tests/src/spl/memo.ts
@@ -1,5 +1,5 @@
 import assert from "assert";
-import { splMemoProgram } from "@project-serum/spl-memo";
+import { splMemoProgram } from "@coral-xyz/spl-memo";
 import { Keypair } from "@solana/web3.js";
 
 import { SPL_MEMO_PROGRAM_ID } from "../constants";